ai/ai_back/requirements.txt
fastapi
uvicorn

pyaudio
obs-websocket-py
websockets
python-dotenv
